get all currency prices

+++ app/src/main/java/com/example/currencyapp/model/response/latest/Rates.kt
@@ -2,7 +2,7 @@ package com.example.currencyapp.model.response.latest
 
 data class Rates(
     val EGP: Double,
-    val EUR: Int,
+    val EUR: Double,
     val GBP: Double,
     val USD: Double
 )
